requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Helmet==0.1.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-json-logger==2.0.7 